Move latest-optimistic-slots over to blockstore subcommand
